% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Example function hello world}
\usage{
hello_world(person)
}
\arguments{
\item{person}{A string variable for the person to say hello to.}
}
\description{
Example function hello world
}
\examples{
hello_world("Johnny")
}
