% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test data}
\format{
## `test_data`
A data frame with 10 rows and 2 columns:
\describe{
  \item{x}{Random variable}
  \item{y}{Random variable}
}
 @export
}
\usage{
test_data
}
\description{
An example of how test data is documented. See https://r-pkgs.org/data.html#sec-documenting-data
}
\keyword{datasets}
